change int to long for solving compile error's (void * -> long) convert in linux 64 os

+++ src/RedisResponse.h
@@ -31,7 +31,7 @@
 
 namespace coconut {
 
-typedef int ticket_t;
+typedef long ticket_t;
 
 class RedisResponseImpl;
